chore: refactored large parts of the popup/home component, storage, context and data layer updates
