+++ src/components/hooks/useSelectedUrl.tsx
@@ -1,0 +1,6 @@
+import { getSelectedUrl, saveSelectedUrl } from '../../background/syncStorage';
+import { useFetchData } from './useFetchData';
+
+export const useSelectedUrl = () => {
+  return useFetchData<string>('', getSelectedUrl, saveSelectedUrl);
+};
