+++ src/components/predeployedAccounts/predeployedAccounts.tsx
@@ -1,23 +1,20 @@
 import React, { useEffect } from 'react';
-import { Box, Button, Container, Stack, Typography } from '@mui/material';
-import { ChevronLeft } from '@mui/icons-material';
+import { Box, Button, Stack, Typography } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import { darkTheme } from '../..';
 import {
-  sendMessageToGetUrl,
   sendMessageToSetSelectedAccount,
   sendMessageToSetUrlList,
 } from '../utils/sendMessageBackground';
 import { AccountData } from '../context/interfaces';
 import { useSharedState } from '../context/context';
+import { darkTheme } from '../..';
 
 export const PredeployedAccounts: React.FC = () => {
   const context = useSharedState();
   const {
     accounts,
     setAccounts,
-    url,
-    setUrl,
+    selectedUrl: url,
     devnetIsAlive,
     setDevnetIsAlive,
     selectedAccount,
@@ -27,6 +24,8 @@ export const PredeployedAccounts: React.FC = () => {
     setUrlList,
     configData,
     setConfigData,
+    lastFetchedUrl,
+    setLastFetchedUrl,
   } = context;
   const navigate = useNavigate();
 
@@ -73,17 +72,13 @@ export const PredeployedAccounts: React.FC = () => {
   }
 
   useEffect(() => {
-    sendMessageToGetUrl(setUrl);
-  }, []);
-
-  useEffect(() => {
+    if (!url || url === lastFetchedUrl) return;
     const fetchData = async () => {
       await fetchDataAndPrintAccounts();
+      setLastFetchedUrl(url);
     };
-    if (url) {
-      fetchData();
-    }
-  }, [url]);
+    fetchData();
+  }, [url, lastFetchedUrl, accounts, devnetIsAlive]);
 
   const handleAccountClick = async (clickedAddress: string) => {
     const clickedAccount = accounts.find((account) => account.address === clickedAddress);
@@ -119,10 +114,6 @@ export const PredeployedAccounts: React.FC = () => {
     fetchSelectedAccount();
   }, [selectedAccount]);
 
-  const handleBack = () => {
-    navigate('/');
-  };
-
   const shortenAddress = (address: string, startCount = 12, endCount = 12) =>
     `${address.slice(0, startCount)}...${address.slice(-endCount)}`;
 
@@ -135,26 +126,6 @@ export const PredeployedAccounts: React.FC = () => {
     <>
       {devnetIsAlive && accounts.length > 0 && (
         <section>
-          <Stack direction={'row'} justifyContent={'center'} position={'relative'}>
-            <Box position={'absolute'} top={0} left={0}>
-              <Button
-                size="small"
-                variant={'text'}
-                startIcon={<ChevronLeft />}
-                onClick={handleBack}
-                sx={{
-                  padding: '8px 10px',
-                }}
-              >
-                Back
-              </Button>
-            </Box>
-            <Container>
-              <Typography variant="h6" margin={0} marginY={2}>
-                Accounts
-              </Typography>
-            </Container>
-          </Stack>
           <Stack marginBottom={1}>
             {accounts.map((account, index) => (
               <Box key={index}>
