+++ src/page/Screen1.tsx
@@ -6,37 +6,38 @@ import { useSharedState } from '../components/context/context';
 import { Component } from '../components/context/enum';
 
 export const Screen1 = () => {
-  const context = useSharedState();
-  const { selectedComponent, setSelectedComponent } = context;
+  // const context = useSharedState();
+  // const { selectedComponent, setSelectedComponent } = context;
 
-  window.addEventListener('message', (event) => {
-    if (event.data.type === 'CHANGE_SELECTED_COMPONENT') {
-      const { selectedComponent: component } = event.data;
-      setSelectedComponent(component);
-    }
-  });
+  return <div></div>;
+  // window.addEventListener('message', (event) => {
+  //   if (event.data.type === 'CHANGE_SELECTED_COMPONENT') {
+  //     const { selectedComponent: component } = event.data;
+  //     setSelectedComponent(component);
+  //   }
+  // });
 
-  return (
-    <section>
-      {!selectedComponent && (
-        <div>
-          <button
-            style={{ marginRight: '10px' }}
-            onClick={() => setSelectedComponent(Component.CommandGenerator)}
-          >
-            Docker Command Generator
-          </button>
-          <button onClick={() => setSelectedComponent(Component.DockerRegister)}>
-            Register Running Docker
-          </button>
-        </div>
-      )}
-      {selectedComponent === Component.CommandGenerator && <DockerCommandGenerator />}
-      {selectedComponent === Component.DockerRegister && <RegisterRunningDocker />}
-      <title>Display Accounts</title>
-      <PredeployedAccounts />
-    </section>
-  );
+  // return (
+  //   <section>
+  //     {!selectedComponent && (
+  //       <div>
+  //         <button
+  //           style={{ marginRight: '10px' }}
+  //           onClick={() => setSelectedComponent(Component.CommandGenerator)}
+  //         >
+  //           Docker Command Generator
+  //         </button>
+  //         <button onClick={() => setSelectedComponent(Component.DockerRegister)}>
+  //           Register Running Docker
+  //         </button>
+  //       </div>
+  //     )}
+  //     {selectedComponent === Component.CommandGenerator && <DockerCommandGenerator />}
+  //     {selectedComponent === Component.DockerRegister && <RegisterRunningDocker />}
+  //     <title>Display Accounts</title>
+  //     <PredeployedAccounts />
+  //   </section>
+  // );
 };
 
 export default Screen1;
